# openapi.json
{"openapi":"3.1.0","info":{"title":"Stratogen Backend","version":"Master v1.1"},"paths":{"/chat":{"post":{"summary":"Chat With Ai","operationId":"chat_with_ai_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders":{"post":{"summary":"Place Order","operationId":"place_order_orders_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/risk/state":{"get":{"summary":"Risk State","operationId":"risk_state_risk_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/risk/update":{"post":{"summary":"Risk Update","operationId":"risk_update_risk_update_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/risk/presets":{"get":{"summary":"Risk Presets","operationId":"risk_presets_risk_presets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/risk/preset":{"post":{"summary":"Risk Apply Preset","operationId":"risk_apply_preset_risk_preset_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyPreset"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/risk/schedule":{"get":{"summary":"Risk Get Schedule","operationId":"risk_get_schedule_risk_schedule_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Risk Set Schedule","operationId":"risk_set_schedule_risk_schedule_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/logs":{"get":{"summary":"Get Logs","operationId":"get_logs_logs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":5000,"minimum":1,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drive/folders":{"post":{"summary":"Create Folder","operationId":"create_folder_drive_folders_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFolderReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drive/list":{"get":{"summary":"List Files","operationId":"list_files_drive_list_get","parameters":[{"name":"parentId","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parentid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drive/search":{"get":{"summary":"Search Files","operationId":"search_files_drive_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drive/uploadBase64":{"post":{"summary":"Upload Base64","operationId":"upload_base64_drive_uploadBase64_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadBase64Req"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/send":{"post":{"summary":"Send Email","operationId":"send_email_gmail_send_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GmailMessage"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/sendAttachment":{"post":{"summary":"Send With Attachment","operationId":"send_with_attachment_gmail_sendAttachment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GmailSendWithAttachment"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/labels":{"get":{"summary":"List Labels","operationId":"list_labels_gmail_labels_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Label","operationId":"create_label_gmail_labels_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GmailLabel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/search":{"get":{"summary":"Search","operationId":"search_gmail_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/threads":{"get":{"summary":"List Threads","operationId":"list_threads_gmail_threads_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/messages/{msg_id}":{"get":{"summary":"Get Message","operationId":"get_message_gmail_messages__msg_id__get","parameters":[{"name":"msg_id","in":"path","required":true,"schema":{"type":"string","title":"Msg Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/drafts":{"post":{"summary":"Create Draft","operationId":"create_draft_gmail_drafts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GmailDraft"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gmail/drafts/send":{"post":{"summary":"Send Draft","operationId":"send_draft_gmail_drafts_send_post","parameters":[{"name":"draft_id","in":"query","required":true,"schema":{"type":"string","title":"Draft Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alpaca/clock":{"get":{"tags":["alpaca"],"summary":"Clock","operationId":"clock_api_alpaca_clock_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/alpaca/account":{"get":{"tags":["alpaca"],"summary":"Account","operationId":"account_api_alpaca_account_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/alpaca/debug":{"get":{"tags":["alpaca"],"summary":"Debug","operationId":"debug_api_alpaca_debug_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/alpaca/asset/{symbol}":{"get":{"tags":["alpaca"],"summary":"Asset","operationId":"asset_api_alpaca_asset__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alpaca/assets/validate":{"post":{"tags":["alpaca"],"summary":"Assets Validate","operationId":"assets_validate_api_alpaca_assets_validate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SymbolsIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bridge/ping":{"post":{"summary":"Ping Bridge","operationId":"ping_bridge_bridge_ping_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/system/health":{"get":{"summary":"System Health","operationId":"system_health_system_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response System Health System Health Get"}}}}}}},"/system/config":{"get":{"summary":"System Config","operationId":"system_config_system_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response System Config System Config Get"}}}}}}},"/debug/env":{"get":{"summary":"Read Env","operationId":"read_env_debug_env_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ApplyPreset":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"ApplyPreset"},"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"ChatRequest"},"CreateFolderReq":{"properties":{"name":{"type":"string","title":"Name"},"parentId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parentid"}},"type":"object","required":["name"],"title":"CreateFolderReq"},"GmailAttachment":{"properties":{"filename":{"type":"string","title":"Filename"},"mime_type":{"type":"string","title":"Mime Type"},"data_base64":{"type":"string","title":"Data Base64"}},"type":"object","required":["filename","mime_type","data_base64"],"title":"GmailAttachment"},"GmailDraft":{"properties":{"to":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"To"},"subject":{"type":"string","title":"Subject"},"body":{"type":"string","title":"Body"},"from_addr":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"From Addr"}},"type":"object","required":["subject","body"],"title":"GmailDraft"},"GmailLabel":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"GmailLabel"},"GmailMessage":{"properties":{"to":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"To"},"subject":{"type":"string","title":"Subject"},"body":{"type":"string","title":"Body"},"from_addr":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"From Addr"}},"type":"object","required":["subject","body"],"title":"GmailMessage"},"GmailSendWithAttachment":{"properties":{"to":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"To"},"subject":{"type":"string","title":"Subject"},"body":{"type":"string","title":"Body"},"attachments":{"items":{"$ref":"#/components/schemas/GmailAttachment"},"type":"array","title":"Attachments","default":[]},"from_addr":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"From Addr"}},"type":"object","required":["subject","body"],"title":"GmailSendWithAttachment"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrderRequest":{"properties":{"symbol":{"type":"string","title":"Symbol"},"side":{"type":"string","title":"Side"},"qty":{"type":"number","title":"Qty"},"order_type":{"type":"string","title":"Order Type","default":"market"},"time_in_force":{"type":"string","title":"Time In Force","default":"gtc"},"limit_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Limit Price"},"stop_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Stop Price"},"approx_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Approx Price"}},"type":"object","required":["symbol","side","qty"],"title":"OrderRequest"},"RiskUpdate":{"properties":{"SESSION_ACTIVE":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Session Active"},"COOL_OFF_ACTIVE":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Cool Off Active"},"ORDER_THROTTLE_SECONDS":{"anyOf":[{"type":"integer","maximum":3600.0,"minimum":0.0},{"type":"null"}],"title":"Order Throttle Seconds"},"DAILY_LOSS_LIMIT":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Daily Loss Limit"},"MAX_POSITION_RISK":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Max Position Risk"}},"type":"object","title":"RiskUpdate"},"ScheduleUpdate":{"properties":{"timezone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timezone"},"trading_days":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Trading Days"},"open_local":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Open Local"},"close_local":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Close Local"},"auto_stop":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Auto Stop"},"enabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Enabled"}},"type":"object","title":"ScheduleUpdate"},"SymbolsIn":{"properties":{"symbols":{"items":{"type":"string"},"type":"array","title":"Symbols","description":"Symbols to validate"}},"type":"object","title":"SymbolsIn"},"UploadBase64Req":{"properties":{"name":{"type":"string","title":"Name"},"data_base64":{"type":"string","title":"Data Base64"},"mime_type":{"type":"string","title":"Mime Type","default":"application/octet-stream"},"folderId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Folderid"}},"type":"object","required":["name","data_base64"],"title":"UploadBase64Req"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}